[focus] Implement manual focus mode.
The camera, while claiming to support manual appears to ignore it.  This
sets a flag to indicate manual mode, and sets the real mode to auto.  If
the flag is set, dont focus when pressing the shutter.  Reest the flag
when the mode is changed.

Add manual mode even if camera/claims to not support it

Dont require libaudioresource for building